+++ src/ui/components/base/debug.js
@@ -1,7 +1,7 @@
-const { ProvidableComponent } = require('./provider.js')
+const { ProvidingComponent } = require('./provider.js')
 const { rect, line, point } = require('../../shapes')
 
-class DebuggedComponent extends ProvidableComponent {
+class DebuggedComponent extends ProvidingComponent {
 	_outlineColor = 'rgba(255, 0, 0, 0.4)'
 	_outlineWidth = 1
 	_gapColor = 'rgba(255, 0, 255, 0.3)'