+++ src/ui/components/base/hooks.js
@@ -46,9 +46,9 @@ class HookedComponent {
 	}
 
 	useResolution() {
-		const { lastW, lastH } = this.usePage()
+		const page = this.usePage()
 
-		return { w: lastW, h: lastH }
+		return page.useResolution()
 	}
 
 	usePayload() {